rag-chatbot/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
ollama>=0.1.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.5.0
pypdf>=3.17.0
unstructured>=0.10.0
tabulate>=0.9.0
python-magic>=0.4.27